commit 20-12
Reduced the size of the islands, boats and cannons.
Added a sphere collider so spawning boats will be safe.

TODO -
reduce starting rotation of spawned boats so we can decrease the size of the sphere collider (the sphere mesh is a placeholder, to be replaced with the boat model, the collider will stay even after).

Decide what to do with the rocks