Bugfix : SpawnPointCountArray didn't reset to zero (see description)
BUGFIX :

- BoatCannonBallCollider changed such that it will change directions only if BOTH collided gameObjects have the tag "BoatAfterHit".
Before that, a sailing boat would trigger ChangeDirection() as well, because there are always *two* boats that trigger OnTriggerEnter(), each in its own context.

This bug was "silent" because a sailing boat would change target only if a cannonball colided with it (Doesn't change into damagedMovement() and RotateTowards a new target cause the tag doesn't change to "BoatAfterHit" - see Update() )

- Added a boolean flag that indicates if the ChangeDirection() was casued by a cannonball or another damaged boat, and variable "t" that saves the original *CURRENT* target.

If it was hit by a cannonball, then the *CURRENT* target is originalTarget. Else, it was already hit by a boat before, so its *CURRENT* target is actually nextTarget.

The random nextTarget algorithm remains the same- But instead of "originalTarget", it's based on "t".

GAMEPLAY DECISION :

- Need to decide what exactly happends if a damaged boat, hits a sailing boat?

I would say sailing boat moves the same to its originalTarget, and the damaged boat triggers ChangeDirection().

The bug fix caused that the collider simply *ignores* this scenario (damage boat "goes over" sailing boat - no colliding happends), but since now we know the source of the bug, if we want to handle this scenario, we might need to modify Tags or modify the conditional statements in BoatCannonBallCollider.